implemented deleteWorkflowResults API, now using it to clean up results when backups associated to workflowId deleted, finally fixed issue in sorting causing backups to be deleted out-of-order

+++ src/engine/plugins/pluginUtil/helper.go
@@ -6,6 +6,7 @@ import (
 	"regexp"
 	"engine/util"
 	"path/filepath"
+	"sort"
 )
 
 func ExistsPath(path string) bool {
@@ -37,51 +38,64 @@ func ListBackups(path string) []util.Backup {
 	}
 
 	var backups []util.Backup
-	re := regexp.MustCompile(`(\S+)_(\S+)_(\S+)`)
-    for _, f := range files {
+	type timeSlice []util.Backup
+
+	re := regexp.MustCompile(`(\S+)_(\S+)_(\S+)_(\S+)`)
+  for _, f := range files {
 		var backup util.Backup
 		match := re.FindStringSubmatch(f.Name())
+
 		if len(match) != 0 {
 			backup.Name = match[1]
 			backup.Policy = match[2]
-			backup.Epoch = match[3]
+			backup.WorkflowId = match[3]
 
-			timestamp := util.ConvertEpoch(match[3])
+			epoch := util.StringToInt(match[4])
+			backup.Epoch = epoch
+
+			timestamp := util.ConvertEpoch(match[4])
 			backup.Timestamp = timestamp
+
 			backups = append(backups, backup)
 		}	
 	}
+
+	sort.Sort(util.ByEpoch(backups))
+
 	return backups
 }
 
 func RecursiveDirDelete(dir string) {
-	d, err := os.Open(dir)
+	if ExistsPath(dir) == true {
+		d, err := os.Open(dir)
 
-	if err != nil {
-		LogErrorMessage(err.Error())
-		os.Exit(1)
-	}
-	defer d.Close()
+		if err != nil {
+			LogErrorMessage(err.Error())
+			os.Exit(1)
+		}
+		defer d.Close()
 
-	names, err := d.Readdirnames(-1)
-	if err != nil {
-		LogErrorMessage(err.Error())
-		os.Exit(1)
-	}
+		names, err := d.Readdirnames(-1)
+		if err != nil {
+			LogErrorMessage(err.Error())
+			os.Exit(1)
+		}
 
-	for _, name := range names {
+		for _, name := range names {
 			err = os.RemoveAll(filepath.Join(dir, name))
 			if err != nil {
 				LogErrorMessage(err.Error())
 				os.Exit(1)
 			}
-	}
+		}
 
-	err = os.Remove(dir)
-	if err != nil {
-		LogErrorMessage(err.Error())
-		os.Exit(1)
-	}
+		err = os.Remove(dir)
+		if err != nil {
+			LogErrorMessage(err.Error())
+			os.Exit(1)
+		}
+		LogInfoMessage("Removed directory " + dir + " completed successfully")
+	}	
 }
 
 func ReverseBackupList(backups []util.Backup) chan util.Backup {
